Ayyyy
So I have pictures of numbers ready to go, but they aren’t implemented
yet. I also started work on the border collision, and have made it so
you can loop around on the bottom purple sidewalk, but nowhere else.
 I will have to add the loop to the other sidewalk, and remove the
collision from the river area. I also need to complete the score
thingy/levels, as well as a timer
{[(Score = levels completed. Every level, the speed is randomized, and
the timer is reset every time you complete a level. The timer will get
shorter and shorter(by fractions of a second)]}