asdf
So, the health counter is being implemented, and I have numbers ready
for the health counter. Next class I will have to edit out the numbers,
and prepare the code for displaying them, as well as work out some
little bugs in the respawn system. Eventually, I might also add the
level system.