Images
All the images are prepped for the game. Maybe need to iron out a few
white spots and fuzzy edges in Gimp, but they’re ready. Next time, need
to fix those spots/fuzzy edges, then I need to actually integrate the
images into the life counter system, and add code for the displaying of
the images.